XMLConfigPreProcessorTest extracted from XMLConfigBuilderTest and some minor refactoring made to XmlConfigPreProcessor to make methods shorter
